d/p: Ubuntu Pro, Use a grid to properly align dotted list elements
We were listing the elements using manual formatting and this was broken
for lower resolutions or in general for longer texts, making a dotted
list not properly padded.

So use a grid to manage this, this indeed changes the text used by
translations but because of lp:2068639 it's not a problem because no
translations are available so far.

Also in this way RTL languages are properly handled

The alternative would have been using TextBuffer tags (that didn't break
translations, but it's more annoying to use and it wouldn't fix RTL
languages.

LP: #2062971