Adding EULA page for OEMs
The EULA page is conditionaly displayed if a EULA.pdf file is present under
/usr/share/desktop-provision/eula/.
It will first attempt to load the EULA matching your current locale
(EULA_<locale>.pdf), using the variants first and then fallbacking to base
language if the current varient is not available.
If no match is found, EULA.pdf is then used.

Note that multi-page pdf's are supported.

Co-authored-by: Didier Roche-Tolomelli <didrocks@ubuntu.com>