Allow setting hostname to local account page
Allow an user to edit their hostname on the local account creation page.
The validation for setting a hostname re-uses the same logic from the username
validation to not introduce new strings for translation, while additionally
ensuring the hostname is not already in use on the network.
When an user enters their username, a suggested hostname is syncronised based
on the username, however it can be overidden.

Co-authored-by: Didier Roche-Tolomelli <didrocks@ubuntu.com>