Provide distributor-logo icon
distributor-logo is not a standard icon name, which may not be available
on the chosen icon theme. Indeed, since ubuntu-desktop is not seeding Humanity
anymore, this icon would be missing.

This is installed in hicolor rather than being bundled as a gresources, because
it is used by the desktop file that ubuntu overrides in ubuntu-settings.