a11y: Notify bad username
When the username is invalid, a text label is updated with the
problem. Unfortunately, the content is not notified to the user.

This patch checks for changes in the label, and uses
gtk_accessible_announce() to read out loud the text.

Forwarding: https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/278
Bug-Ubuntu: https://bugs.launchpad.net/ubuntu/+source/gnome-initial-setup/+bug/2116935