Fix SVGs
A change in GTK (https://gitlab.gnome.org/GNOME/gtk/-/commit/416b89af98b181eba25cd4bfa6300c5dbd7db3dd)
broke SVG pictures that lack the width and height properties in
the svg tag.

This patch fixes this by adding them.

Also, Gtk.Image is more oriented for icons, and with that patch,
the last image, the one about the app store, was painted
incorrectly. This is also fixed by changing it to a Gtk.Picture.

Finally, the size of the pictures was smaller, probably due to a
different pixel-size default value. Also fixed here.